move some configuration enums to a more logical place
